requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
idna==3.2
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dotenv==0.19.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
